169 - Dropping support for Android 4.4 (#236)
Ticket:
https://github.com/medic/cht-android/issues/169

In this commit:
- Removes the support for Android 4.4.
- Refactor the WebView's url handler and adds support for sdk 21 & 22 (Android 5)
- Removes android.support (deprecated) and use the new one: androidx.
