+++ .github/workflows/publish.yml
@@ -263,11 +263,8 @@ jobs:
       if: startsWith(github.ref, 'refs/tags/')
       with:
         draft: true
-        # APKs for webview-armeabi-v7a are not uploaded
         files: |
-          build/outputs/apk/**/*-xwalk-arm64-v8a-release.apk
-          build/outputs/apk/**/*-xwalk-armeabi-v7a-release.apk
-          build/outputs/apk/**/*-webview-arm64-v8a-release.apk
+          build/outputs/apk/**/*-release.apk
           build/outputs/bundle/**/*-release.aab
       env:
         GITHUB_TOKEN: ${{ secrets.RELEASE_TOKEN }}
