+++ Makefile
@@ -1,7 +1,7 @@
 ADB = ${ANDROID_HOME}/platform-tools/adb
 GRADLE = ./gradlew
 GRADLE_OPTS = --daemon --parallel
-flavor = UnbrandedWebview
+flavor = Unbranded
 abi = x86
 KEYTOOL = keytool
 OPENSSL = openssl
@@ -21,21 +21,20 @@ ifdef ComSpec	 # Windows
 endif
 
 default: deploy logs
-xwalk: deploy-xwalk logs
 
 logs:
 	${ADB} logcat MedicMobile:V AndroidRuntime:E '*:S' | tee android.log
 
 deploy:
 	${GRADLE} ${GRADLE_OPTS} install${flavor}Debug
-deploy-xwalk:
-	${GRADLE} ${GRADLE_OPTS} installUnbrandedXwalkDebug
+
 deploy-all:
 	find build/outputs/apk -name \*-debug.apk | \
 		xargs -n1 ${ADB} install -r
 
 clean:
 	${GRADLE} clean
+
 clean-apks:
 	rm -rf build/outputs/apk/
 
@@ -43,10 +42,12 @@ assemble: check-env
 	ANDROID_KEYSTORE_PATH=${ANDROID_KEYSTORE_PATH} ANDROID_KEY_ALIAS=${ANDROID_KEY_ALIAS} \
 	ANDROID_KEYSTORE_PASSWORD=${ANDROID_KEYSTORE_PASSWORD} ANDROID_KEY_PASSWORD=${ANDROID_KEY_PASSWORD} \
 	        ${GRADLE} ${GRADLE_OPTS} assemble${flavor}
+
 assemble-all: check-env
 	ANDROID_KEYSTORE_PATH=${ANDROID_KEYSTORE_PATH} ANDROID_KEY_ALIAS=${ANDROID_KEY_ALIAS} \
 	ANDROID_KEYSTORE_PASSWORD=${ANDROID_KEYSTORE_PASSWORD} ANDROID_KEY_PASSWORD=${ANDROID_KEY_PASSWORD} \
 	        ${GRADLE} ${GRADLE_OPTS} assembleRelease
+
 assemble-all-debug:
 	${GRADLE} ${GRADLE_OPTS} assembleDebug
 
@@ -54,6 +55,7 @@ bundle: check-env
 	ANDROID_KEYSTORE_PATH=${ANDROID_KEYSTORE_PATH} ANDROID_KEY_ALIAS=${ANDROID_KEY_ALIAS} \
 	ANDROID_KEYSTORE_PASSWORD=${ANDROID_KEYSTORE_PASSWORD} ANDROID_KEY_PASSWORD=${ANDROID_KEY_PASSWORD} \
 	        ${GRADLE} ${GRADLE_OPTS} bundle${flavor}Release
+
 bundle-all: check-env
 	ANDROID_KEYSTORE_PATH=${ANDROID_KEYSTORE_PATH} ANDROID_KEY_ALIAS=${ANDROID_KEY_ALIAS} \
 	ANDROID_KEYSTORE_PASSWORD=${ANDROID_KEYSTORE_PASSWORD} ANDROID_KEY_PASSWORD=${ANDROID_KEY_PASSWORD} \
@@ -70,19 +72,23 @@ uninstall:
 
 lint:
 	${GRADLE} ${GRADLE_OPTS} androidCheck lint${flavor}Debug
+
 test: lint
 	${GRADLE} ${GRADLE_OPTS} test
 
 test-ui:
-	${GRADLE} connectedUnbrandedWebviewDebugAndroidTest \
+	${GRADLE} connectedUnbrandedDebugAndroidTest \
 		-Pabi=${abi} --stacktrace -Pandroid.testInstrumentationRunnerArguments.class=\
 	org.medicmobile.webapp.mobile.SettingsDialogActivityTest
+
 test-ui-gamma:
-	${GRADLE} connectedMedicmobilegammaWebviewDebugAndroidTest -Pabi=${abi} --stacktrace
+	${GRADLE} connectedMedicmobilegammaDebugAndroidTest -Pabi=${abi} --stacktrace
+
 test-ui-url:
-	DISABLE_APP_URL_VALIDATION=true ${GRADLE} connectedUnbrandedWebviewDebugAndroidTest \
+	DISABLE_APP_URL_VALIDATION=true ${GRADLE} connectedUnbrandedDebugAndroidTest \
 		-Pabi=${abi} --stacktrace -Pandroid.testInstrumentationRunnerArguments.class=\
 	org.medicmobile.webapp.mobile.LastUrlTest
+
 test-ui-all: test-ui test-ui-gamma test-ui-url
 
 test-bash-keystore:
