+++ fastlane/Fastfile
@@ -6,13 +6,11 @@ default_platform(:android)
 
 platform :android do
   lane :build do |options|
-    gradle(task: "assemble#{options[:flavor]}WebviewRelease")
-    gradle(task: "assemble#{options[:flavor]}XwalkRelease")
+    gradle(task: "assemble#{options[:flavor]}Release")
   end
 
   lane :bundle do |options|
-    gradle(task: "bundle#{options[:flavor]}WebviewRelease")
-    gradle(task: "bundle#{options[:flavor]}XwalkRelease")
+    gradle(task: "bundle#{options[:flavor]}Release")
   end
 
   lane :deploy do |options|
@@ -24,11 +22,9 @@ platform :android do
       track: "alpha",
       json_key: "playstore-secret.json",
       apk_paths: [
-        "build/outputs/apk/#{options[:flavor]}Webview/release/cht-android-#{version}-#{options[:flavor]}-webview-arm64-v8a-release.apk",
-        # NOT SUPPORTED WITH: targetSdkVersion 29. Need to add it back to support older Android versions
-        # "build/outputs/apk/#{options[:flavor]}Webview/release/cht-android-#{version}-#{options[:flavor]}-webview-armeabi-v7a-release.apk",
-        "build/outputs/apk/#{options[:flavor]}Xwalk/release/cht-android-#{version}-#{options[:flavor]}-xwalk-arm64-v8a-release.apk",
-        "build/outputs/apk/#{options[:flavor]}Xwalk/release/cht-android-#{version}-#{options[:flavor]}-xwalk-armeabi-v7a-release.apk",
+        "build/outputs/apk/#{options[:flavor]}/release/cht-android-#{version}-#{options[:flavor]}-arm64-v8a-release.apk",
+        # Support for older versions of Android:
+        "build/outputs/apk/#{options[:flavor]}/release/cht-android-#{version}-#{options[:flavor]}-armeabi-v7a-release.apk",
       ],
       skip_upload_aab: true,
       skip_upload_metadata: true,
@@ -48,8 +44,7 @@ platform :android do
       track: "alpha",
       json_key: "playstore-secret.json",
       aab_paths: [
-        "build/outputs/bundle/#{options[:flavor]}WebviewRelease/cht-android-#{version}-#{options[:flavor]}-webview-release.aab",
-        "build/outputs/bundle/#{options[:flavor]}XwalkRelease/cht-android-#{version}-#{options[:flavor]}-xwalk-release.aab",
+        "build/outputs/bundle/#{options[:flavor]}Release/cht-android-#{version}-#{options[:flavor]}-release.aab",
       ],
       skip_upload_apk: true,
       skip_upload_metadata: true,
