+++ gradle.properties
@@ -1,2 +1,4 @@
 android.useAndroidX=true
+# Remove 'android.enableJetifier=true' when dropping the libraries: com.simprints:LibSimprints and com.github.Mariovc:ImagePicker.
+android.enableJetifier=true
 org.gradle.jvmargs=-Xmx2048m
