--- src/main/assets/xwalk-command-line
@@ -1,1 +1,0 @@
-xwalk --unlimited-storage
