+++ src/main/res/layout/main.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+		android:id="@+id/lytWebView"
+		android:layout_width="fill_parent"
+		android:layout_height="fill_parent">
+	<WebView android:id="@+id/wbvMain"
+			android:layout_width="fill_parent"
+			android:layout_height="fill_parent"/>
+</LinearLayout>
