--- src/webview/AndroidManifest.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="org.medicmobile.webapp.mobile">
-
-  <application android:label="@string/app_name"
-      android:icon="@mipmap/ic_launcher"
-      android:allowBackup="false"
-      android:largeHeap="true"
-      tools:ignore="GoogleAppIndexingWarning,LockedOrientationActivity,Instantiatable">
-    <activity android:name="UpgradingActivity"
-        android:screenOrientation="portrait"
-        android:configChanges="orientation|screenSize"
-        android:launchMode="singleTop"/>
-  </application>
-
-</manifest>
