--- src/xwalk/res/layout/main.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-		android:id="@+id/lytWebView"
-		android:layout_width="fill_parent"
-		android:layout_height="fill_parent">
-	<org.xwalk.core.XWalkView android:id="@+id/wbvMain"
-			android:layout_width="fill_parent"
-			android:layout_height="fill_parent"/>
-</LinearLayout>
